pyproject.toml
[tool.poetry]
name = "prbot-py"
version = "0.1.0"
description = ""
authors = ["Denis BOURGE <bourge.denis@gmail.com>"]
readme = "README.md"
packages = [
    { include = "prbot" }
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
inject = "^5.2.1"
typer = "^0.12.3"
tortoise-orm = {extras = ["asyncpg"], version = "^0.21.5"}
orjson = "^3.10.3"
aerich = "^0.7.2"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
httpx = "^0.27.0"
pydantic-settings = "^2.3.1"
structlog = "^24.2.0"
backoff = "^2.2.1"
sentry-sdk = {extras = ["fastapi"], version = "^2.7.1"}
redis = "^5.0.6"
python-dotenv = "^1.0.1"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
cryptography = "^42.0.8"
prometheus-client = "^0.20.0"
prometheus-fastapi-instrumentator = "^7.0.0"
gunicorn = "^22.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.8"
pytest = "^8.2.2"
mypy = "^1.10.0"
debugpy = "^1.8.1"
pytest-cov = "^5.0.0"
isort = "^5.13.2"

[tool.mypy]
disable_error_code = "var-annotated,type-abstract"

[tool.aerich]
tortoise_orm = "prbot.modules.database.settings.TORTOISE_ORM"
location = "./prbot/modules/database/migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
